-finds locationID -newer c920 cameras did not seem to work with UVC code on macOS  >= 10.12 -that's fixed now

+++ src/ofxUVCUseProber.h
@@ -9,7 +9,8 @@ void useCameraByDeviceIndex(int i){
     
     ofLog()<<"index "<<i<<", vendorID "<<vendorID<<" productID "<<productID<<" locationID "<<locationID;
 //    useCamera(vendorID,productID,0x00);
-    useCamera2(vendorID,productID,0x00,(int)locationID);
+//    useCamera2(vendorID,productID,0x00,(int)locationID);
+      useCamera2(vendorID,productID,0x00,locationID);
 }
 void useFirstCamera(){
     useCameraByDeviceIndex(0);
